requirements.txt
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.2.1
click==8.1.8
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.104.1
h11==0.14.0
idna==3.10
passlib==1.7.4
pyasn1==0.6.1
pydantic==2.10.5
pydantic_core==2.27.2
PyJWT==2.8.0
python-decouple==3.8
python-jose==3.3.0
python-multipart==0.0.20
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.12.2
uvicorn==0.24.0
